+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Kraken parser: added new export format. ([#469](https://github.com/BittyTax/BittyTax/issues/469))
 - sFOX parser: added tramsaction history parser.
 - Kraken parser: added "earn" subtypes "airdrop" and "autoallocation".
+- SwissBorg parser: added "Fee Adjustment" Type.
 ### Changed
 - Price tool: CoinDesk API deprecated.